requirements.txt
aiohttp==1.0.5
appdirs==1.4.3
async-timeout==1.2.0
chardet==2.3.0
discord.py==0.16.7
multidict==2.1.4
packaging==16.8
pexpect==4.2.1
ptyprocess==0.5.1
pyparsing==2.2.0
six==1.10.0
websockets==3.3
